+++ CHANGES.md
@@ -14,6 +14,11 @@
 - add error handling around shorthand non-boolean attributes
 - add support for shorthand boolean attributes
 - read version number from toml
+- source_path in component definition relative to rvo instead of absolute path
+- move component definitions to src/jinja-roos-components/definitions
+- update conversion script and autocomplete generation script to use new paths
+- add `regenerate_components` script to run conversion on all converted components at once
+- add tests for all components, some commented out because currently failing
 
 # [0.1] - 2025-11-04
 - First tagged version